requirements.txt
Django==1.11
djangorestframework==3.11.0
future==0.18.2
pymysql==0.9.3
serializers==0.2.4
